feat: catalog handler 패키지 수정
