+++ apps/commerce-streamer/src/main/kotlin/com/loopers/domain/catalog/handler/ProductLikedEventHandler.kt
@@ -1,4 +1,4 @@
-package com.loopers.infrastructure.catalog
+package com.loopers.domain.catalog.handler
 
 import com.loopers.domain.catalog.CatalogEventHandler
 import com.loopers.event.EventType
@@ -9,18 +9,18 @@ import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Component
 
 @Component
-class ProductUnlikedEventHandler(
+class ProductLikedEventHandler(
     private val cacheRepository: CacheRepository,
 ) : CatalogEventHandler<ProductLikeChangedEvent> {
 
     private val log = LoggerFactory.getLogger(this::class.java)
 
     override fun handle(eventPayload: ProductLikeChangedEvent) {
-        log.info("[ProductUnlikedEventHandler.handle] eventPayload: $eventPayload")
+        log.info("[ProductLikedEventHandler.handle] eventPayload: $eventPayload")
         cacheRepository.evict(CacheNames.PRODUCT_LIKE_COUNT_V1 + eventPayload.productId)
     }
 
     override fun supports(eventType: EventType): Boolean {
-        return EventType.PRODUCT_UNLIKED == eventType
+        return EventType.PRODUCT_LIKED == eventType
     }
 }
